SCHD_relu_leastSquares.py
This is a basic machine learning model that uses 15 ReLU activated layers to determine the next day's high stock price for the SCHD stock. The model uses Mean Squared Error as it's loss function and RMSprop with a learning rate of .001 as its optimization function. The data is not standardized within this model, so the numbers are put into the model raw.