multiFunctionTest.py
When run, this file will pull the data from each of the stocks, normalize the data, then fit the model using all combinations of datasets, activation functions, and optimization. The  mean absolute percent error of each epoch will then be output to an excel file titled modelResults.xls, and the final results (average of all kfolds) will be output to the console with the dataset, activation function, and optimization function specified.