ModelResults 4_5_2023.xlsx, multiFunctionTest.py, Poster.pptx
ModelResults 4_5_2023.xlsx is data from when everything was run in the code. This data was shown on 4/5/2023 during the meeting.
multiFunctionTest.py is code to create and test all models for all combinations of datasets, activations, and optimization functions. The code used to test SGD is commented out due to complications.
Poster.pptx is the draft of the poster shown on 4/5/2023 during the meeting.